+++ game_patch/rf/event.h
@@ -324,7 +324,8 @@ namespace rf
         Set_Item_Drop,
         AF_Heal,
         Anchor_Marker_Orient,
-        Light_State
+        Light_State,
+        World_HUD_Sprite
     };
 
     // int to EventType